+++ app/src/androidTest/java/com/mani/baking/activity/RecipeMainActivityTest.java
@@ -5,7 +5,7 @@ import android.content.res.Configuration;
 
 import com.mani.baking.R;
 import com.mani.baking.datastruct.IngredientDetails;
-import com.mani.baking.datastruct.Recipe;
+import com.mani.baking.utils.SessionData;
 import com.mani.baking.datastruct.RecipeDetails;
 import com.mani.baking.datastruct.StepDetails;
 import com.mani.baking.utils.ExtractJson;
@@ -62,7 +62,7 @@ public class RecipeMainActivityTest {
     public void setUpSessionVar() {
         ExtractJson extractJson = new ExtractJson(InstrumentationRegistry.getInstrumentation().getTargetContext());
         extractJson.initializeSessionVar();
-        recipeDetailsList = Recipe.recipeDetailsList;
+        recipeDetailsList = SessionData.recipeDetailsList;
     }
 
     /**
@@ -159,7 +159,7 @@ public class RecipeMainActivityTest {
             getPrevButton().perform(click());
             getPrevButton().check(isDisabled());
         }
-        List<StepDetails> stepDetailsList = Recipe.getRecipeDetails(i).getStepDetailsList();
+        List<StepDetails> stepDetailsList = SessionData.getRecipeDetails(i).getStepDetailsList();
 
         for (int j = 0; j < stepDetailsList.size(); j++) {
